+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Prevented recursion when setting the value of readable writable properties 
 * Added tooltip text property for JComponent
 * Added foreground color and background color properties for Component
+* Added value property for JFormattedTextField
 * Added convenient comparison method in ValueUtils
 * Added ConstantTransformer to always return a specific value
 * Fixed JToggleButtonSelectedProperty value